# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "消防巡检管理系统 API",
    "description": "基于 Node.js + Express + MongoDB 的消防设备巡检、报警处理、安全管理系统 API 文档。\n\n## 认证方式\n所有私有接口需在请求头中携带 Authorization: Bearer <token>。\n\n## 数据范围\n部分接口受数据范围权限控制：all（全部）/ department（本部门）/ self（仅本人）/ none（无权限）。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "/",
      "description": "当前服务器"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/security/my-info": {
      "get": {
        "tags": ["Security"],
        "summary": "获取当前用户安全信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回安全评分、近期登录记录、安全建议"
          }
        }
      }
    },
    "/api/security/change-password": {
      "put": {
        "tags": ["Security"],
        "summary": "修改密码（强化版，需确认密码）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentPassword", "newPassword", "confirmPassword"],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功，需重新登录"
          }
        }
      }
    },
    "/api/security/bindings": {
      "get": {
        "tags": ["Security"],
        "summary": "获取账户绑定信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回邮箱、手机、部门绑定状态（脱敏）"
          }
        }
      }
    },
    "/api/security/view-sensitive": {
      "post": {
        "tags": ["Security"],
        "summary": "查看敏感数据（需二次验证）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dataType"],
                "properties": {
                  "dataType": {
                    "type": "string",
                    "enum": ["phone", "email"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "返回脱敏和完整数据"
          }
        }
      }
    },
    "/api/security/stats": {
      "get": {
        "tags": ["Security"],
        "summary": "获取系统安全统计（管理员）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回今日登录数、失败登录数、高风险操作数、异常行为"
          }
        }
      }
    },
    "/api/security/report": {
      "post": {
        "tags": ["Security"],
        "summary": "举报异常行为",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targetType", "reason"],
                "properties": {
                  "targetType": {
                    "type": "string",
                    "enum": ["user", "device", "alarm", "system"]
                  },
                  "targetId": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "举报已提交"
          }
        }
      }
    },
    "/api/security/my-logs": {
      "get": {
        "tags": ["Security"],
        "summary": "获取个人操作日志",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 7
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/security/users/{userId}/lock": {
      "put": {
        "tags": ["Security"],
        "summary": "锁定/解锁用户账户（管理员）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["locked"],
                "properties": {
                  "locked": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "403": {
            "description": "无权限"
          }
        }
      }
    },
    "/api/security/overview": {
      "get": {
        "tags": ["Security"],
        "summary": "获取安全概览（管理员）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回严重/高危告警数、失败登录数、风险评分"
          }
        }
      }
    },
    "/api/security/alerts": {
      "get": {
        "tags": ["Security"],
        "summary": "获取最近安全告警",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回最近50条告警"
          }
        }
      }
    },
    "/api/security/audit-logs": {
      "get": {
        "tags": ["Security"],
        "summary": "查询审计日志（多维度筛选）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 1000
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ip",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "riskLevel",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          },
          {
            "name": "success",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["info", "warning", "error"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/security/config/allowPublicRegistration": {
      "get": {
        "tags": ["Security"],
        "summary": "获取注册开关状态",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回 allowPublicRegistration"
          }
        }
      },
      "put": {
        "tags": ["Security"],
        "summary": "设置注册开关",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["allowPublicRegistration"],
                "properties": {
                  "allowPublicRegistration": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/security/config/loginCaptchaEnabled": {
      "get": {
        "tags": ["Security"],
        "summary": "获取登录验证码开关状态",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回 loginCaptchaEnabled"
          }
        }
      },
      "put": {
        "tags": ["Security"],
        "summary": "设置登录验证码开关",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["loginCaptchaEnabled"],
                "properties": {
                  "loginCaptchaEnabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/security/ip-list": {
      "get": {
        "tags": ["Security"],
        "summary": "获取 IP 黑白名单列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["black", "white"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Security"],
        "summary": "添加 IP 到黑/白名单",
        "description": "白名单优先级高于黑名单；加入黑名单时若已在白名单则拒绝",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ip"],
                "properties": {
                  "ip": {
                    "type": "string",
                    "description": "支持 IPv4/IPv6/CIDR"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["black", "white"],
                    "default": "black"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "durationHours": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 8760,
                    "description": "0=永久"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "添加成功"
          },
          "400": {
            "description": "IP 已在白名单中"
          }
        }
      }
    },
    "/api/security/ip-list/{id}": {
      "delete": {
        "tags": ["Security"],
        "summary": "从名单中移除 IP 记录",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/security/ip-list/query": {
      "get": {
        "tags": ["Security"],
        "summary": "查询 IP 命中的黑/白名单记录",
        "description": "返回该 IP 命中的全部黑/白名单记录（含 CIDR 网段与等价地址形式）；各列表按覆盖面最宽优先排序（如 /16 与 /24 同时命中时 /16 在首位），并给出最终生效判定（白名单优先）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "IPv4/IPv6 单地址（不含 CIDR）"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回命中记录"
          }
        }
      }
    },
    "/api/security/users/{userId}/mfa/reset": {
      "put": {
        "tags": ["Security"],
        "summary": "管理员重置用户两步验证（清除密钥/恢复码并强制下线）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "两步验证已重置"
          },
          "400": {
            "description": "未开启 MFA / 目标为自身"
          },
          "403": {
            "description": "同级或更高层级 / 内置超管"
          }
        }
      }
    },
    "/api/security/audit-logs/verify": {
      "get": {
        "tags": ["Security"],
        "summary": "校验审计日志哈希链完整性（hash 重算 + hmac + 链接性三层校验）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回校验报告（intact/breaks/byType）"
          }
        }
      }
    },
    "/api/security/audit-logs/export": {
      "get": {
        "tags": ["Security"],
        "summary": "导出审计日志（CSV + sha256 签名 manifest，流式输出）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/csv 流式响应"
          }
        }
      }
    },
    "/api/security/config/registerCaptchaEnabled": {
      "get": {
        "tags": ["Security"],
        "summary": "获取注册验证码开关状态",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "返回 { registerCaptchaEnabled: boolean }"
          }
        }
      },
      "put": {
        "tags": ["Security"],
        "summary": "切换注册验证码开关",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["registerCaptchaEnabled"],
                "properties": {
                  "registerCaptchaEnabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "开关已更新"
          }
        }
      }
    },
...
  }
}